make the demo code better
That means we will show first the very needed code (which is only the `find`, everything else has sane defaults). a link under `more options` goes to the documentation when js enabled (well it should anyway for holmes to work, but PE anyway 🎉) or shows all of the options identical to the README. Don't forget to update here as well in case of a change.

fixes #34

cc @lukyvj who noticed this.

screenshots

closed|open
---|---
![](https://i.imgur.com/SzZYW3v.png)|![](https://i.imgur.com/SmS5a3G.png)