clean up implementation of placeholder hide/show
fixes #78

so the problem mostly was that the piece of code that showed and hid the placeholder became quite complex when it was moved around and when the callbacks were added. Another reason was that I used negated ifs, which made it harder for myself. Anyway, completely rewritten now, as well as gave the demo the callbacks